+++ com.unity.render-pipelines.universal/Runtime/ForwardRendererData.cs
@@ -70,7 +70,9 @@ namespace UnityEngine.Rendering.Universal
         public ShaderResources shaders = null;
 
         [SerializeField] LayerMask m_OpaqueLayerMask = -1;
+        [SerializeField] uint m_OpaqueRenderingLayerMask = uint.MaxValue;
         [SerializeField] LayerMask m_TransparentLayerMask = -1;
+        [SerializeField] uint m_TransparentRenderingLayerMask = uint.MaxValue;
         [SerializeField] StencilStateData m_DefaultStencilState = new StencilStateData() { passOperation = StencilOp.Replace }; // This default state is compatible with deferred renderer.
         [SerializeField] bool m_ShadowTransparentReceive = true;
         [SerializeField] RenderingMode m_RenderingMode = RenderingMode.Forward;
@@ -106,6 +108,19 @@ namespace UnityEngine.Rendering.Universal
         }
 
         /// <summary>
+        /// Use this to configure how to filter opaque objects by rendering layer.
+        /// </summary>
+        public uint opaqueRenderingLayerMask
+        {
+            get => m_OpaqueRenderingLayerMask;
+            set
+            {
+                SetDirty();
+                m_OpaqueRenderingLayerMask = value;
+            }
+        }
+
+        /// <summary>
         /// Use this to configure how to filter transparent objects.
         /// </summary>
         public LayerMask transparentLayerMask
@@ -115,6 +130,19 @@ namespace UnityEngine.Rendering.Universal
             {
                 SetDirty();
                 m_TransparentLayerMask = value;
+            }
+        }
+
+        /// <summary>
+        /// Use this to configure how to filter transparent objects by rendering layer.
+        /// </summary>
+        public uint transparentRenderingLayerMask
+        {
+            get => m_TransparentRenderingLayerMask;
+            set
+            {
+                SetDirty();
+                m_TransparentRenderingLayerMask = value;
             }
         }