+++ com.unity.render-pipelines.universal/Runtime/Passes/DepthNormalOnlyPass.cs
@@ -18,9 +18,9 @@ namespace UnityEngine.Rendering.Universal.Internal
         /// <summary>
         /// Create the DepthNormalOnlyPass
         /// </summary>
-        public DepthNormalOnlyPass(RenderPassEvent evt, RenderQueueRange renderQueueRange, LayerMask layerMask)
+        public DepthNormalOnlyPass(RenderPassEvent evt, RenderQueueRange renderQueueRange, LayerMask layerMask, uint renderingLayerMask = uint.MaxValue)
         {
-            m_FilteringSettings = new FilteringSettings(renderQueueRange, layerMask);
+            m_FilteringSettings = new FilteringSettings(renderQueueRange, layerMask, renderingLayerMask);
             renderPassEvent = evt;
         }