+++ com.unity.render-pipelines.universal/Runtime/Passes/DepthOnlyPass.cs
@@ -21,9 +21,9 @@ namespace UnityEngine.Rendering.Universal.Internal
         /// <summary>
         /// Create the DepthOnlyPass
         /// </summary>
-        public DepthOnlyPass(RenderPassEvent evt, RenderQueueRange renderQueueRange, LayerMask layerMask)
+        public DepthOnlyPass(RenderPassEvent evt, RenderQueueRange renderQueueRange, LayerMask layerMask, uint renderingLayerMask = uint.MaxValue)
         {
-            m_FilteringSettings = new FilteringSettings(renderQueueRange, layerMask);
+            m_FilteringSettings = new FilteringSettings(renderQueueRange, layerMask, renderingLayerMask);
             renderPassEvent = evt;
         }