+++ com.unity.render-pipelines.universal/Runtime/Passes/GBufferPass.cs
@@ -24,12 +24,12 @@ namespace UnityEngine.Rendering.Universal.Internal
         FilteringSettings m_FilteringSettings;
         RenderStateBlock m_RenderStateBlock;
 
-        public GBufferPass(RenderPassEvent evt, RenderQueueRange renderQueueRange, LayerMask layerMask, StencilState stencilState, int stencilReference, DeferredLights deferredLights)
+        public GBufferPass(RenderPassEvent evt, RenderQueueRange renderQueueRange, LayerMask layerMask, StencilState stencilState, int stencilReference, DeferredLights deferredLights, uint renderingLayerMask = uint.MaxValue)
         {
             base.renderPassEvent = evt;
 
             m_DeferredLights = deferredLights;
-            m_FilteringSettings = new FilteringSettings(renderQueueRange, layerMask);
+            m_FilteringSettings = new FilteringSettings(renderQueueRange, layerMask, renderingLayerMask);
             m_RenderStateBlock = new RenderStateBlock(RenderStateMask.Nothing);
 
             m_RenderStateBlock.stencilState = stencilState;